+++ test/libsolidity/syntaxTests/inheritance/override/override_shared_base_partial.sol
@@ -8,3 +8,5 @@ contract B is I {}
 contract C is A, B {
   function f() external override(A, I) {}
 }
+// ----
+// Warning: (178-192): Override specifier list order differs from inheritance order.