+++ test/libsolidity/syntaxTests/nameAndTypeResolution/055_inheritance_diamond_basic.sol
@@ -3,6 +3,6 @@ contract inter1 is root { function f() public virtual {} }
 contract inter2 is root { function f() public virtual {} }
 contract derived is root, inter2, inter1 {
 	function g() public { f(); rootFunction(); }
-	function f() override(inter1, inter2) public {}
+	function f() override(inter2, inter1) public {}
 }
 // ----